backend/tickets/requirements.txt
psycopg2
PyJWT
requests
